ai_trader_npc_fastapi/requirements.txt
fastapi
uvicorn
web3
python-dotenv
google-generativeai
eth_account
gunicorn
